Use standardized exposure value between 0 and 1

+++ android/src/main/java/com/google/android/cameraview/CameraViewImpl.java
@@ -76,9 +76,9 @@ abstract class CameraViewImpl {
 
     abstract int getFlash();
 
-    abstract void setExposureCompensation(int exposure);
+    abstract void setExposureCompensation(float exposure);
 
-    abstract int getExposureCompensation();
+    abstract float getExposureCompensation();
 
     abstract void takePicture(ReadableMap options);
